CC_App_Backend/CC_App_Backend/cricket_coach_app/requirements.txt.txt
absl-py==2.1.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
idna==3.7
intel-openmp==2021.4.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mediapipe==0.10.14
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.44
scipy==1.14.0
seaborn==0.13.2
setuptools==72.1.0
six==1.16.0
sounddevice==0.4.7
sqlparse==0.5.1
sympy==1.13.1
tbb==2021.13.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.63
ultralytics-thop==2.0.0
urllib3==2.2.2
wheel==0.43.0